Rename package towards feature-based packaging
