+++ appholder/src/main/java/com/android/mdl/app/documentdata/RequestMicovAtt.kt
@@ -1,8 +1,9 @@
-package com.android.mdl.app.request
+package com.android.mdl.app.documentdata
 
 import com.android.mdl.app.R
 
 object RequestMicovAtt : RequestDocument() {
+
     override val docType = "org.micov.1"
     override val nameSpace = "org.micov.attestation.1"
     override val dataItems = DataItems.values().asList()
